requirements.txt
jsonpath_ng==1.6.1
PyMySQL==1.0.2
pytest==7.4.3
PyYAML==5.4.1
PyYAML==6.0.2
Requests==2.32.3
xlrd==1.2.0
